The code starts with making activity and calling them based on the pressed button. To call and activity first you need to make an empty activity in by RC on the java folder. The next step is to make onClickListener objects. the onClickListener objects are overwriting the default onClick method.
and OnClickListener object is a subclass of View, so to make the object you need to call it as a child of View:
View.OnClickListener nameOftheOnClickListener= new View.OnClickListener (){
@Override
    public void onClick (View v){
        Intent activityIntent= new Intent (SourceActivity.This, GoalActivity.class)
        startActivity(activityIntent);
    }
}

and overwrite the onClick by generating an intent to open the goal activity from the source activity.
In this code instead of generating 4 different onClickListener object I only did four if/else to check the iD of the view using getView().